Rename internal model and minor code clean up.
